Fix kernel panic error and system hang in adaptive streaming
:Release Notes:
Fix kernel panic error and system hang in adaptive streaming

:Detailed Notes:
With the changes done for restore saved crtc mode the saved_crtc
is not freed due to the check of modesetting_enabled flag when
restore_saved_crtc is called. Rearanged the code to free it.

But disabled the complete saved crtc code since it is applicable
only for force mode setting which is false by default

Again when drmModeObjectSetProperty is set for 0xff04 then dest
values of scale_param are not initialized, making it provide junk
values (very big numbers) sometimes.
Since the dest is not expected and it is expected to be set by
set display window call so initialized them to 0 (zero)

:Testing Performed:
Tested adaptive streaming playback using youtube test app
and also in youtube desktop site using browser

:QA Notes:

:Issues Addressed:
[PLAT-62115] [OSE] Display resetting in Videosink during adaptive
             streaming

Change-Id: I27ca77b3dfb2a6acb3d08415ae10a8a29a82e56d
Reviewed-on: http://gpro.lgsvl.com/210665
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Biju Malayil <biju.malayil@lge.com>
Tested-by: Biju Malayil <biju.malayil@lge.com>